refactor: CoinCOunt를 외부로 이동
